+++ relay-server/src/services/processor.rs
@@ -45,9 +45,7 @@ use tokio::sync::Semaphore;
 use {
     crate::services::store::{Store, StoreEnvelope},
     crate::utils::{EnvelopeLimiter, ItemAction, MetricsLimiter},
-    relay_cardinality::{
-        CardinalityLimit, CardinalityLimiter, CardinalityScope, RedisSetLimiter, SlidingWindow,
-    },
+    relay_cardinality::{CardinalityLimit, CardinalityLimiter, RedisSetLimiter},
     relay_metrics::{Aggregator, RedisMetricMetaStore},
     relay_quotas::{RateLimitingError, RedisRateLimiter},
     relay_redis::RedisPool,
@@ -123,6 +121,8 @@ pub enum ProcessingGroup {
     CheckIn,
     /// Spans.
     Span,
+    /// Metrics.
+    Metrics,
     /// Unknown item types will be forwarded upstream (to processing Relay), where we will
     /// decide what to do with them.
     ForwardUnknown,
@@ -458,15 +458,6 @@ impl<'a> ProcessEnvelopeState<'a> {
     /// Returns a mutable reference to the contained [`Envelope`].
     fn envelope_mut(&mut self) -> &mut Envelope {
         self.managed_envelope.envelope_mut()
-    }
-
-    /// Returns whether any item in the envelope creates an event in any relay.
-    ///
-    /// This is used to branch into the processing pipeline. If this function returns false, only
-    /// rate limits are executed. If this function returns true, an event is created either in the
-    /// current relay or in an upstream processing relay.
-    fn creates_event(&self) -> bool {
-        self.envelope().items().any(Item::creates_event)
     }
 
     /// Returns true if there is an event in the processing state.
@@ -827,34 +818,6 @@ impl EnvelopeProcessorService {
         }
     }
 
-    /// Returns current cardinality limits built from the configuration.
-    ///
-    /// In the near future these limits will be coming from the project config.
-    #[cfg(feature = "processing")]
-    fn get_cardinality_limits(&self) -> [CardinalityLimit; 5] {
-        let limit = self.inner.config.cardinality_limit();
-        let window = SlidingWindow {
-            window_seconds: self.inner.config.cardinality_limiter_window(),
-            granularity_seconds: self.inner.config.cardinality_limiter_granularity(),
-        };
-        let scope = CardinalityScope::Organization;
-
-        [
-            MetricNamespace::Sessions,
-            MetricNamespace::Spans,
-            MetricNamespace::Transactions,
-            MetricNamespace::Custom,
-            MetricNamespace::Unsupported,
-        ]
-        .map(|ns| CardinalityLimit {
-            id: ns.as_str().to_owned(),
-            window,
-            limit,
-            scope,
-            namespace: Some(ns),
-        })
-    }
-
     /// Normalize monitor check-ins and remove invalid ones.
     #[cfg(feature = "processing")]
     fn process_check_ins(&self, state: &mut ProcessEnvelopeState) {
@@ -1193,7 +1156,11 @@ impl EnvelopeProcessorService {
             self.extract_metrics(state)?;
         }
 
-        dynamic_sampling::sample_envelope_items(state);
+        dynamic_sampling::sample_envelope_items(
+            state,
+            &self.inner.config,
+            &self.inner.global_config.current(),
+        );
 
         if_processing!(self.inner.config, {
             event::store(state, &self.inner.config, self.inner.geoip_lookup.as_ref())?;
@@ -1213,11 +1180,8 @@ impl EnvelopeProcessorService {
         Ok(())
     }
 
-    /// Processes standalone attachments.
-    fn process_standalone_attachments(
-        &self,
-        state: &mut ProcessEnvelopeState,
-    ) -> Result<(), ProcessingError> {
+    /// Processes standalone items that require an event ID, but do not have an event on the same envelope.
+    fn process_standalone(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
         profile::filter(state);
 
         if_processing!(self.inner.config, {
@@ -1288,86 +1252,6 @@ impl EnvelopeProcessorService {
         Ok(())
     }
 
-    /// Legacy implementation of the `process_state` function, which is used for all
-    /// unknown [`ProcessingGroup`] variant.
-    ///
-    /// Note: this will be removed once we are confident in the new implementation and make sure
-    /// that all the groups properly covered.
-    fn process_state_legacy(
-        &self,
-        state: &mut ProcessEnvelopeState,
-    ) -> Result<(), ProcessingError> {
-        session::process(state, &self.inner.config);
-        report::process_client_reports(
-            state,
-            &self.inner.config,
-            self.inner.outcome_aggregator.clone(),
-        );
-        report::process_user_reports(state);
-        replay::process(state, &self.inner.config)?;
-        profile::filter(state);
-        span::filter(state);
-
-        if state.creates_event() {
-            // Some envelopes only create events in processing relays; for example, unreal events.
-            // This makes it possible to get in this code block while not really having an event in
-            // the envelope.
-
-            if_processing!(self.inner.config, {
-                unreal::expand(state, &self.inner.config)?;
-            });
-
-            event::extract(state, &self.inner.config)?;
-            profile::transfer_id(state);
-
-            if_processing!(self.inner.config, {
-                unreal::process(state)?;
-                attachment::create_placeholders(state);
-            });
-
-            event::finalize(state, &self.inner.config)?;
-            self.light_normalize_event(state)?;
-            dynamic_sampling::normalize(state);
-            event::filter(state)?;
-            dynamic_sampling::run(state, &self.inner.config);
-
-            // We avoid extracting metrics if we are not sampling the event while in non-processing
-            // relays, in order to synchronize rate limits on indexed and processed transactions.
-            if self.inner.config.processing_enabled() || state.sampling_result.should_drop() {
-                self.extract_metrics(state)?;
-            }
-
-            dynamic_sampling::sample_envelope_items(state);
-
-            if_processing!(self.inner.config, {
-                event::store(state, &self.inner.config, self.inner.geoip_lookup.as_ref())?;
-            });
-        }
-
-        if_processing!(self.inner.config, {
-            self.enforce_quotas(state)?;
-            profile::process(state, &self.inner.config);
-            self.process_check_ins(state);
-            span::process(
-                state,
-                self.inner.config.clone(),
-                self.inner.global_config.current().measurements.as_ref(),
-            );
-        });
-
-        if state.has_event() {
-            event::scrub(state)?;
-            event::serialize(state)?;
-            if_processing!(self.inner.config, {
-                span::extract_from_event(state);
-            });
-        }
-
-        attachment::scrub(state);
-
-        Ok(())
-    }
-
     fn process_state(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
         // Get the group from the managed envelope context, and if it's not set, try to guess it
         // from the contents of the envelope.
@@ -1379,21 +1263,26 @@ impl EnvelopeProcessorService {
             ProcessingGroup::Error => self.process_errors(state)?,
             ProcessingGroup::Transaction => self.process_transactions(state)?,
             ProcessingGroup::Session => self.process_sessions(state)?,
-            ProcessingGroup::Standalone => self.process_standalone_attachments(state)?,
+            ProcessingGroup::Standalone => self.process_standalone(state)?,
             ProcessingGroup::ClientReport => self.process_client_reports(state)?,
             ProcessingGroup::Replay => self.process_replays(state)?,
             ProcessingGroup::CheckIn => self.process_checkins(state)?,
             ProcessingGroup::Span => self.process_spans(state)?,
+            // Currently is not used.
+            ProcessingGroup::Metrics => {
+                relay_log::error!(
+                    tags.project = %state.project_id,
+                    items = ?state.envelope().items().next().map(Item::ty),
+                    "received metrics in the process_state"
+                );
+            }
             // Fallback to the legacy process_state implementation for Ungrouped events.
             ProcessingGroup::Ungrouped => {
                 relay_log::error!(
                     tags.project = %state.project_id,
                     items = ?state.envelope().items().next().map(Item::ty),
-                    "Could not identify the processing group based on the envelope's items"
+                    "could not identify the processing group based on the envelope's items"
                 );
-
-                // Call the legacy implementation of the `process_state` function.
-                self.process_state_legacy(state)?;
             }
             // Leave this group unchanged.
             //
@@ -1717,6 +1606,7 @@ impl EnvelopeProcessorService {
             envelope,
             self.inner.outcome_aggregator.clone(),
             self.inner.test_store.clone(),
+            ProcessingGroup::ClientReport,
         );
         self.handle_submit_envelope(SubmitEnvelope { envelope });
     }
@@ -1879,8 +1769,6 @@ impl EnvelopeProcessorService {
         use crate::constants::DEFAULT_EVENT_RETENTION;
         use crate::services::store::StoreMetrics;
 
-        let limits = &self.get_cardinality_limits();
-
         for (scoping, message) in message.scopes {
             let ProjectMetrics {
                 mut buckets,
@@ -1888,6 +1776,7 @@ impl EnvelopeProcessorService {
             } = message;
 
             let mode = project_state.get_extraction_mode();
+            let limits = project_state.get_cardinality_limits();
 
             if project_state.has_feature(Feature::CardinalityLimiter) {
                 buckets = self.cardinality_limit_buckets(scoping, limits, buckets, mode);
@@ -1968,6 +1857,7 @@ impl EnvelopeProcessorService {
                         envelope,
                         self.inner.outcome_aggregator.clone(),
                         self.inner.test_store.clone(),
+                        ProcessingGroup::Metrics,
                     );
                     envelope.set_partition_key(partition_key).scope(scoping);
 
@@ -2107,6 +1997,7 @@ impl EnvelopeProcessorService {
             envelope,
             self.inner.outcome_aggregator.clone(),
             self.inner.test_store.clone(),
+            ProcessingGroup::Metrics,
         );
         self.handle_submit_envelope(SubmitEnvelope { envelope });
     }
@@ -2639,8 +2530,15 @@ mod tests {
 
         let mut project_state = ProjectState::allowed();
         project_state.config = config;
+
+        let mut envelopes = ProcessingGroup::split_envelope(*envelope);
+        assert_eq!(envelopes.len(), 1);
+
+        let (group, envelope) = envelopes.pop().unwrap();
+        let envelope = ManagedEnvelope::standalone(envelope, outcome_aggregator, test_store, group);
+
         let message = ProcessEnvelope {
-            envelope: ManagedEnvelope::standalone(envelope, outcome_aggregator, test_store),
+            envelope,
             project_state: Arc::new(project_state),
             sampling_project_state: None,
             reservoir_counters: ReservoirCounters::default(),